+++ RELEASE_NOTES.md
@@ -34,6 +34,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - Dome info panels now also include 
     - Altitude if available
     - Info if following is off, following via N.I.N.A. or following via driver to differentiate the different dome following mechanisms
+- Options > Astrometry can now show a world map with a pin of your entered latitude and longitude
 
 ## Commandline Options
 The following command line options have been added